// editor/src/utils/api.ts
export interface FileDto {
  /** @format guid */
  id: string;
  path?: string;
  isImageFromSocial: boolean;
  fullPath?: string;
  originalFileName?: string;
  fullPathWebP?: string;
  fullPathJp2?: string;
  dominantColor?: string;
  type: FileType;

  /** @format int32 */
  width: number;

  /** @format int32 */
  height: number;

  /** @format double */
  size: number;
}

export enum FileType {
  Favicon = 'Favicon',
  SocialImage = 'SocialImage',
  Preview = 'Preview',
  TemplatePreview = 'TemplatePreview',
  TemplateFull = 'TemplateFull',
  ProfileAvatar = 'ProfileAvatar',
  ProfileCover = 'ProfileCover',
  Elements = 'Elements',
  Community = 'Community'
}